refactor: Extracted project navigation into it's own component
